# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A user system API starter.",
        "title": "Frame",
        "contact": {
            "name": "Amir Irani",
            "email": "freshmanlimited@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3500",
    "basePath": "/endpoint",
    "paths": {
        "/admin/auth/admin-manage/create": {
            "post": {
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adminManage"
                ],
                "summary": "Create a new Admin",
                "parameters": [
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.Message"
                        }
                    }
                }
            }
        },
        "/admin/auth/admin-manage/get/all": {
            "get": {
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adminManage"
                ],
                "summary": "Get all admins",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.Message"
                        }
                    }
                }
            }
        },
        "/admin/auth/admin-manage/get/{id}": {
            "get": {
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adminManage"
                ],
                "summary": "Get admin by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "adminID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.Message"
                        }
                    }
                }
            }
        },
        "/admin/auth/admin-manage/password/{id}": {
            "put": {
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adminManage"
                ],
                "summary": "Update password of admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "adminID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.Message"
                        }
                    }
                }
            }
        },
        "/admin/auth/admin-manage/status/{id}": {
            "put": {
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adminManage"
                ],
                "summary": "Set admin status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "adminID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "IsActive",
                        "name": "isActive",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.Message"
                        }
                    }
                }
            }
        },
        "/admin/auth/admin-manage/username/{id}": {
            "put": {
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adminManage"
                ],
                "summary": "Update username of admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "adminID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.Message"
                        }
                    }
                }
            }
        },
...
        "/admin/signin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Admin signin",
                "parameters": [
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.Message"
                        }
                    }
                }
            }
        },
        "/admin/signout": {
            "delete": {
                "security": [
                    {
                        "AdminApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Admin signout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/response.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.Message": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Message": {
                    "type": "object"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminApiKeyAuth": {
            "type": "apiKey",
            "name": "AdminAuthorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}